Initialize task-planner-server crate
The `task-planner-server` crate will provide the back-end server to
maintain the tasks, plans, and other information for this
application.

We are using the [diesel](https://diesel.rs/) crate to manage SQL
migrations and define data types.

In the future, we may want to share type definitions between the
server and various clients. This may entail splitting a
`task-planner-core` away from this crate. However, at this point, I'm
not sure how to make that work with the traits required by diesel for
SQL interactions.