Create users table
The users table today only contains an ID, a username, and a name. In
the future, this will likely contain other information in order to
allow people to log in. However, I don't want to deal with this
complexity at first. For now, we will simply allow the user to type in
their username to log in without any kind of authentication.