Create add_user binary
Create the necessary models and helper methods to add a new user to
the database. As a bonus step, the binary checks to see if a user
already exists with the given username.

It is possible, of course, that another user could be added with the
same database in between when we check and when we try to add
it. However, the uniqueness constraint on the database will prevent us
from accidentally adding two users with the same username. This will
cause the insert statement to fail. This is an accepted risk. It is
highly unlikely, and if our cli/webpage/whatever displays an error in
situations like this then that's fine.