Check for uniqueness violation when adding user
The first implementation would panic if there was a uniqueness
violation when adding a user. This was changed to do a database query
to check if the username is taken and prompt for a new
username. However, this is susceptible to "time-of-check to
time-of-use" problems.

Instead of simply panicking or displaying an error when the uniqueness
constraint is violated, we can simply check for that specific kind of
error. If a uniqueness constraint is violated, then we can display the
specific "that username is taken" message. If it's any other kind of
error, we can just display an error or panic or whatever.