Add Rocket and r2d2 pooling
In preparation for adding a server, we add dependencies on rocket and
r2d2. Besides adding the simple example server binary, we start using
a connection pool instead of a single connection.

Our current implementation simply establishes a pool and immediately
extracts a single connection, which is not the typical way a
connection pool is used. However, since all of the usages are in
single binaries, there's not much more we could do.

In the future, if these binaries are merged into a simple shell that
keeps running after commands are given, then we could take advantage
of the pooling. However, even then, we probably won't need more than
one connection at a time.

The real benefit will come when we use the server. Although it's not
buying us much right now, it's nice to make one change at a time and
make sure everything is still working.