Display tasks as HTML
Return an actual HTML web page for the `/user/<username>/tasks`
endpoint.

Each task can be "completed" or "uncompleted" by pressing a link. This
means that performing a GET request to a specific endpoint is
sufficient to mark a task as "complete" or "incomplete". This seems a
little weird to me. Should I be using a PUT request? According to
[MDN][forms], apparently forms use either POST or GET (or dialog?),
with GET being the preferred method when the form has no side
effect. I believe that this action has no side effects since the same
thing happens whether you do it once or several times.

I decided to use Tera as my templating engine because I am somewhat
familiar with Jinja2, which Tera is based on.

[forms]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/form#attr-method